backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
rembg==2.0.56
pillow==10.4.0
python-multipart==0.0.9
